9xb Tech Test
Used php artisan make:controller EmployeeController --resource to generate the CRUD controller for employees, got the edit working easily enough without the role drop-down, then added that and the add new employee.

Pretty much copied the code for the roles - decided to put this in it's own mini-section so you can manage roles then assign them to existing/new employees.

Todo: It needs a bit of tidying and testing - need to make sure it works in other browsers too.

Signed-off-by: Neil <neil.whitaker@whichbingo.co.uk>