MVC Epiphany!
It just clicked that the point of models (and the M in MVC) was so that you abstract the data from the controllers, so that it shouldn't matter what format or type of database you are using as long as the model returns the right information.

So I amended the Roles and Employees models and then the corresponding controllers to use these.

It's so much neater now!

Signed-off-by: Neil <neil.whitaker@gmail.com>